lesson10/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Django==5.2
django-dotenv==1.4.2
django-filter==25.1
django-modeltranslation==0.19.14
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.1.0
idna==3.10
inflection==0.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
